refactor: Move function to submodule
Move it to module for GitHub API related tasks.
